Complete dashboard session, editor and identity boundaries
