+++ app/dashboard/machine-selectors.ts
@@ -6,11 +6,13 @@ export function summarizeLiveMachines(
   savedMachines: MachinePlan[] = [],
 ): MachinePlan[] {
   const grouped = new Map<string, MachinePlan>();
-  const addOutput = (list: MachineOutput[] | undefined, name: string) => {
+  const addOutput = (list: MachineOutput[] | undefined, name: string, rawId?: string | null, rawVariant?: string | null) => {
     const outputs = list || [];
-    const existing = outputs.find((item) => item.name === name);
+    const id = rawId ? qualifyItemId(rawId) : undefined;
+    const variant = rawVariant ? qualifyItemId(rawVariant) : undefined;
+    const existing = outputs.find((item) => id ? item.id === id && item.variant === variant : !item.id && item.name === name);
     if (existing) existing.count += 1;
-    else outputs.push({ name, count: 1 });
+    else outputs.push({ ...(id ? { id } : {}), ...(variant ? { variant } : {}), name, count: 1 });
     return outputs;
   };
   for (const item of items) {
@@ -40,11 +42,11 @@ export function summarizeLiveMachines(
     if (!machine.locations!.includes(item.location))
       machine.locations!.push(item.location);
     if (item.output && item.ready)
-      machine.readyOutputs = addOutput(machine.readyOutputs, item.output);
+      machine.readyOutputs = addOutput(machine.readyOutputs, item.output, item.outputId, item.outputVariant);
     else if (item.output && item.processing)
-      machine.workingOutputs = addOutput(machine.workingOutputs, item.output);
+      machine.workingOutputs = addOutput(machine.workingOutputs, item.output, item.outputId, item.outputVariant);
     if (item.input && item.processing)
-      machine.inputs = addOutput(machine.inputs, item.input);
+      machine.inputs = addOutput(machine.inputs, item.input, item.inputId, item.inputVariant);
     if (item.processing && (item.minutesUntilReady || 0) > 0)
       machine.nextReadyMinutes =
         machine.nextReadyMinutes === null
