+++ desktop/resources/bridge/manifest.json
@@ -1,7 +1,7 @@
 {
   "Name": "Maglucen Companion Bridge",
   "Author": "maglucen",
-  "Version": "5.1.0",
+  "Version": "5.1.1",
   "Description": "Exports read-only live state for Maglucen Stardew Valley Companion without adding an in-game interface.",
   "UniqueID": "maglucen.StardewValleyToolBridge",
   "EntryDll": "StardewValleyToolBridge.dll",
